refactor : modify directory structure
