+++ src/App.js
@@ -1,52 +1,41 @@
-import logo from './logo.svg';
-import './App.css';
-import Home from './pages/hom';
-import './pages/main.css';
-import Contact from './pages/contact';
-import About from './pages/about';
-import {Link,Route,BrowserRouter,Switch} from 'react-router-dom';
-import Login from './pages/login';
+import "./style/App.css";
+import "./style/main.css";
 
-import { Navbar,Container,Nav,NavDrop,Button} from 'react-bootstrap';
-import Logins from './pages/logins';
-import Main from './pages/main';
-function Ea(){
-  return(
+import { Link, Route, BrowserRouter, Switch } from "react-router-dom";
+import { Navbar, Container, Nav, NavDrop, Button } from "react-bootstrap";
+
+import Main from "./views/main";
+import Home from "./views/home";
+import About from "./views/about";
+import Contact from "./views/contact";
+import Login from "./views/login/login";
+import Logins from "./views/login/logins";
+
+function Ea() {
+  return (
     <div>
       <h3>hkhk</h3>
     </div>
   );
-
 }
 
 function App() {
   return (
-    <div >
-    
-  
-     <Switch>
+    <div>
+      <Switch>
         <Route exact path="/" component={Ea}>
-        <Logins />
+          <Logins />
         </Route>
 
         <Route exact path="/main">
-         
-         <Main/>
+          <Main />
         </Route>
-     
-        <Route exact  path="/contact">
-         <Login/>
+
+        <Route exact path="/contact">
+          <Login />
         </Route>
-       
-     </Switch>
-
-
-
-
-
-
-  
-      </div>
+      </Switch>
+    </div>
   );
 }
