--- src/pages/main.js
@@ -1,96 +1,0 @@
-
-import React from 'react';
-import './main.css';
-import { Navbar,Container,Nav,NavDrop,Button,Offcanvas,NavDropdown,Form,FormControl} from 'react-bootstrap';
-const Main = () =>{
-
-
-    return (
-        <div>
- <Navbar bg="light" expand={false}>
-  <Container fluid>
-    <Navbar.Brand href="#">
-    Filpplus</Navbar.Brand>
- 
-    <Navbar.Toggle aria-controls="offcanvasNavbar" />
-    <Navbar.Offcanvas
-      id="offcanvasNavbar"
-      aria-labelledby="offcanvasNavbarLabel"
-      placement="end"
-    >
-      <Offcanvas.Header closeButton>
-        <Offcanvas.Title id="offcanvasNavbarLabel">Offcanvas</Offcanvas.Title>
-      </Offcanvas.Header>
-      <Offcanvas.Body>
-        <Nav className="justify-content-end flex-grow-1 pe-3">
-          <Nav.Link href="#action1">Home</Nav.Link>
-          <Nav.Link href="#action2">Link</Nav.Link>
-          <NavDropdown title="Dropdown" id="offcanvasNavbarDropdown">
-            <NavDropdown.Item href="#action3">Action</NavDropdown.Item>
-            <NavDropdown.Item href="#action4">Another action</NavDropdown.Item>
-            <NavDropdown.Divider />
-            <NavDropdown.Item href="#action5">
-              Something else here
-            </NavDropdown.Item>
-          </NavDropdown>
-        </Nav>
-        <Form className="d-flex">
-          <FormControl
-            type="search"
-            placeholder="Search"
-            className="me-2"
-            aria-label="Search"
-          />
-          <Button variant="outline-success">Search</Button>
-        </Form>
-      </Offcanvas.Body>
-    </Navbar.Offcanvas>
-  </Container>
-</Navbar>
-
-
-<Container className='background'>
-
-
-   <div className="garbage">
-       
-   </div>
- </Container>
-   
- <div className="row">
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-   </div>
-   <div className="row">
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-<div className="col-md-3">
-<a href="" className="button">hihi</a>
-</div>
-   </div>
-
-        </div>
-    );
-};
-
-export default Main;
-
-
